refactor(engine): extract daemon I/O helpers to src/daemon/io/
Five I/O functions extracted from workflow-runner.ts to focused modules:
- io/soul-loader.ts: loadDaemonSoul, WORKRAIL_DIR
- io/workspace-context-loader.ts: loadWorkspaceContext, stripFrontmatter, workspace constants
- io/session-notes-loader.ts: loadSessionNotes, MAX_SESSION_RECAP_NOTES, MAX_SESSION_NOTE_CHARS
- io/conversation-log.ts: appendConversationMessages
- io/execution-stats.ts: writeExecutionStats, writeStuckOutboxEntry, DAEMON_STATS_DIR

New architecture test enforces io/ does not import SDK clients or session-management
modules (@anthropic-ai/sdk, @anthropic-ai/bedrock-sdk, session-scope.ts,
active-sessions.ts). Type imports from agent-loop.ts are allowed
(appendConversationMessages needs AgentInternalMessage).

Also addresses MR review findings F1-F4:
- F1: Extended NODE_IMPORT_PATTERN to cover events, stream, buffer, url, assert, timers
- F2: Added state->core import direction test (state/ must not depend on core/)
- F4: Added 13 unit tests for named state transitions in session-state-transitions.test.ts

378/378 test files, 5858 tests pass (pre-existing ENOTEMPTY flakiness unrelated).
Build clean.
